Fix indent and split component
